+++ src/main/java/roomescape/system/auth/dto/LoginCheckResponse.java
@@ -1,4 +1,4 @@
-package roomescape.auth.dto;
+package roomescape.system.auth.dto;
 
 public record LoginCheckResponse(
         String name
