+++ src/main/java/roomescape/system/auth/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package roomescape.auth.dto;
+package roomescape.system.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
