+++ src/test/java/roomescape/auth/service/AuthServiceTest.java
@@ -6,9 +6,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
-import roomescape.auth.dto.LoginRequest;
+import roomescape.system.auth.dto.LoginRequest;
 import roomescape.system.auth.jwt.JwtHandler;
 import roomescape.system.auth.jwt.dto.TokenDto;
+import roomescape.system.auth.service.AuthService;
 import roomescape.system.exception.model.NotFoundException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
